Organize react components in web/static/js/components in dockup_ui
