+++ apps/dockup_ui/web/static/js/app.js
@@ -19,4 +19,3 @@ import "phoenix_html"
 // paths "./socket" or full ones "web/static/js/socket".
 
 // import socket from "./socket"
-import "web/static/js/index"
