+++ apps/dockup_ui/web/static/js/components/deployment_display.js
