+++ apps/dockup_ui/web/static/js/index.js
@@ -1,4 +1,4 @@
 import React from "react"
 import ReactDOM from "react-dom"
-import DeploymentForm from "web/static/js/deployment_form"
+import DeploymentForm from "web/static/js/components/deployment_form"
 ReactDOM.render(<DeploymentForm />, document.getElementById('deployments_container'));
